fixes incorrect deletion of wrong shields in some situations

+++ code/game/machinery/shieldgen.dm
@@ -499,7 +499,8 @@
 		T2 = T
 		if(locate(/obj/machinery/shieldwall) in T)
 			F = (locate(/obj/machinery/shieldwall) in T)
-			qdel(F)
+			if(F.type == /obj/machinery/shieldwall)
+				qdel(F)
 
 		if(locate(/obj/machinery/shieldwallgen) in T)
 			G = (locate(/obj/machinery/shieldwallgen) in T)
